testes/4-teste_de_api_e_interface_web/backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
psycopg2-binary==2.9.9
python-dotenv==1.0.0
pydantic==2.5.0
pydantic-settings==2.1.0
SQLAlchemy==2.0.23
cachetools==5.3.2
